# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "hamidreza87m@gmail.com"
        },
        "version": "8.0.0"
    },
    "paths": {
        "/api/discount_codes": {
            "get": {
                "tags": [
                    "DiscountCodes"
                ],
                "summary": "Display a listing of discount code.",
                "description": "Get All DiscountCodes.",
                "operationId": "GetAllDiscountCodes",
                "responses": {
                    "200": {
                        "description": "All Discount Codes response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountCodes"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DiscountCodes"
                ],
                "summary": "Store a new Discount Code.",
                "description": "Create a new DiscountCode.",
                "operationId": "CreateDiscountCode",
                "requestBody": {
                    "description": "Code to add",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/DiscountCodeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Discount response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountCode"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/discount_codes/{discount_code_id}": {
            "get": {
                "tags": [
                    "DiscountCodes"
                ],
                "summary": "Display the specified resource.",
                "description": "Get DiscountCode.",
                "operationId": "GetDiscountCode",
                "parameters": [
                    {
                        "name": "discount_code_id",
                        "in": "path",
                        "description": "DiscountCode id which is going to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DiscountCodes response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountCodes"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/discount_histories/use_discount": {
            "post": {
                "tags": [
                    "DiscountHistories"
                ],
                "summary": "Use discount code.",
                "description": "Deposit DiscountHistory.",
                "operationId": "DepositDiscountHistory",
                "requestBody": {
                    "description": "Its needed to create user and DiscountCode at first.",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/DiscountHistoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DiscountHistory response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DiscountHistory"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display a listing of users resource.",
                "description": "Get All Users.",
                "operationId": "GetAllUsers",
                "responses": {
                    "200": {
                        "description": "All users response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store a new User.",
                "description": "Create a new User.",
                "operationId": "CreateUser",
                "requestBody": {
                    "description": "Just enter the mobile below",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/users/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display the user resource.",
                "description": "Get User.",
                "operationId": "GetUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id which is going to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/wallets/deposit": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Deposit amount to a wallet.",
                "description": "Deposit a users wallet.",
                "operationId": "DepositWallet",
                "requestBody": {
                    "description": "Deposit a users wallet",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/WalletRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wallet response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wallet"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "DiscountCodeRequest": {
                "required": [
                    "code"
                ],
                "properties": {
                    "code": {
                        "description": "New disctount code",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DiscountHistoryRequest": {
                "required": [
                    "discount_code_id",
                    "user_id"
                ],
                "properties": {
                    "discount_code_id": {
                        "description": "DiscountCode id",
                        "type": "int"
                    },
                    "user_id": {
                        "description": "User id",
                        "type": "int"
                    }
                },
                "type": "object"
            },
            "UserRequest": {
                "required": [
                    "mobile"
                ],
                "properties": {
                    "mobile": {
                        "description": "Enter a number at least 10 min length",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "WalletRequest": {
                "required": [
                    "deposit",
                    "user_id"
                ],
                "properties": {
                    "deposit": {
                        "description": "How many you whant to apply",
                        "type": "int"
                    },
                    "user_id": {
                        "description": "Enter the selected user id",
                        "type": "int"
                    }
                },
                "type": "object"
            },
            "DiscountCode": {
                "properties": {
                    "id": {
                        "description": "Class DiscountCodeResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "description": "Class DiscountCodeResource",
                        "type": "string"
                    },
                    "number_used": {
                        "description": "Class DiscountCodeResource",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
            "DiscountCodes": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/DiscountCode"
                }
            },
            "DiscountHistory": {
                "properties": {
                    "id": {
                        "description": "Class DiscountHistoryResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "discount_code_id": {
                        "description": "Class DiscountHistoryResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "Class DiscountHistoryResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "discount_code": {
                        "$ref": "#/components/schemas/DiscountCode"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "DiscountHistories": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/DiscountHistory"
                }
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "Class UserResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "mobile": {
                        "description": "Class UserResource",
                        "type": "string"
                    },
                    "name": {
                        "description": "Class UserResource",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Users": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/User"
                }
            },
            "Wallet": {
                "properties": {
                    "id": {
                        "description": "Class WalletResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "Class WalletResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "deposit": {
                        "description": "Class WalletResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "withdraw": {
                        "description": "Class WalletResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "balance": {
                        "description": "Class WalletResource",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Wallets": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Wallet"
                }
            }
        }
    },
    "security": [
        []
    ]
}